# packages/api-client/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/posts/{slug}": {
      "get": {
        "operationId": "posts_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            },
            "example": "arquitetura-aplicacoes-nestjs"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "308": {
            "description": "O slug informado é histórico; redireciona para o slug atual.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Post inexistente ou não publicado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "INTERNAL_ERROR",
                  "details": null,
                  "message": "Erro interno do servidor.",
                  "path": "/api/v1/posts/{slug}",
                  "requestId": "019c2d62-6e90-7000-8000-000000000000",
                  "statusCode": 500,
                  "timestamp": "2026-08-27T20:15:00.000Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "Falha interna inesperada.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Consulta um post publicado pelo slug",
        "tags": ["Posts"]
      }
    },
...
    "/api/v1/admin/posts": {
      "get": {
        "operationId": "adminPosts_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            },
            "example": 1
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            },
            "example": 20
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PostStatus"
            },
            "example": "ARCHIVED"
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "example": "arquitetura",
              "type": "string"
            },
            "example": "arquitetura"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPostAdminSummaryDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Dados ou parâmetros inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Acesso exclusivo de administrador.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Post não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Estado ou slug incompatível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Conteúdo ou slug semanticamente inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "INTERNAL_ERROR",
                  "details": null,
                  "message": "Erro interno do servidor.",
                  "path": "/api/v1/admin/posts",
                  "requestId": "019c2d62-6e90-7000-8000-000000000000",
                  "statusCode": 500,
                  "timestamp": "2026-08-27T20:15:00.000Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "Falha interna inesperada.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-jwt": []
          }
        ],
        "summary": "Lista posts para administração",
        "tags": ["Admin Posts"]
      },
      "post": {
        "operationId": "adminPosts_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostAdminDetailDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Dados ou parâmetros inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Acesso exclusivo de administrador.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Post não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Estado ou slug incompatível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Conteúdo ou slug semanticamente inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "INTERNAL_ERROR",
                  "details": null,
                  "message": "Erro interno do servidor.",
                  "path": "/api/v1/admin/posts",
                  "requestId": "019c2d62-6e90-7000-8000-000000000000",
                  "statusCode": 500,
                  "timestamp": "2026-08-27T20:15:00.000Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "Falha interna inesperada.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-jwt": []
          }
        ],
        "summary": "Cria um rascunho",
        "tags": ["Admin Posts"]
      }
    },
    "/api/v1/admin/posts/{id}/revisions": {
      "get": {
        "operationId": "adminPosts_listRevisions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            },
            "example": "019c2d62-6e90-7000-8000-000000000010"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            },
            "example": 1
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            },
            "example": 20
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPostRevisionAdminDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Dados ou parâmetros inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Acesso exclusivo de administrador.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Post não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Estado ou slug incompatível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Conteúdo ou slug semanticamente inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "INTERNAL_ERROR",
                  "details": null,
                  "message": "Erro interno do servidor.",
                  "path": "/api/v1/admin/posts/{id}/revisions",
                  "requestId": "019c2d62-6e90-7000-8000-000000000000",
                  "statusCode": 500,
                  "timestamp": "2026-08-27T20:15:00.000Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "Falha interna inesperada.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador usado para correlacionar a requisição nos logs.",
                "schema": {
                  "example": "019c2d62-6e90-7000-8000-000000000000",
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-jwt": []
          }
        ],
        "summary": "Lista o histórico de edições de um post publicado",
        "tags": ["Admin Posts"]
      }
    },
...
  },
  "info": {
    "title": "Vavito Archives API",
    "description": "Contrato HTTP da API do Vavito Archives. Corpos JSON e URL-encoded aceitam no máximo 1 MiB; uploads multipart possuem limites próprios por rota.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token do Supabase Auth.",
        "type": "http"
      }
    },
    "schemas": {
...
      "PostDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019c2d62-6e90-7000-8000-000000000010",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "arquitetura-aplicacoes-nestjs"
          },
          "title": {
            "type": "string",
            "example": "Arquitetura de aplicações NestJS"
          },
          "excerpt": {
            "type": "string",
            "example": "Uma introdução prática à arquitetura do projeto."
          },
          "coverUrl": {
            "type": "string",
            "example": "https://cdn.example.com/posts/capa.webp",
            "nullable": true
          },
          "coverAlt": {
            "type": "string",
            "example": "Diagrama de arquitetura",
            "nullable": true
          },
          "coverScale": {
            "type": "number",
            "example": 100,
            "maximum": 160,
            "minimum": 100
          },
          "coverPositionX": {
            "type": "number",
            "example": 50,
            "maximum": 100,
            "minimum": 0
          },
          "coverPositionY": {
            "type": "number",
            "example": 50,
            "maximum": 100,
            "minimum": 0
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagResponseDto"
            }
          },
          "publishedAt": {
            "type": "string",
            "example": "2026-08-20T12:00:00.000Z",
            "format": "date-time"
          },
          "readingTimeMinutes": {
            "type": "number",
            "example": 6,
            "minimum": 0
          },
          "viewCount": {
            "type": "number",
            "example": 128,
            "minimum": 0
          },
          "author": {
            "$ref": "#/components/schemas/PostPublicAuthorDto"
          },
          "content": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "content": [
                {
                  "type": "paragraph"
                }
              ],
              "type": "doc"
            }
          },
          "contentSchemaVersion": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "seoTitle": {
            "type": "string",
            "example": "Arquitetura NestJS",
            "nullable": true
          },
          "seoDescription": {
            "type": "string",
            "example": "Aprenda a organizar uma aplicação NestJS.",
            "nullable": true
          },
          "reactionCounts": {
            "$ref": "#/components/schemas/PostReactionCountsDto"
          },
          "viewer": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PostViewerStateDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "excerpt",
          "coverUrl",
          "coverAlt",
          "tags",
          "publishedAt",
          "readingTimeMinutes",
          "viewCount",
          "author",
          "content",
          "contentSchemaVersion",
          "seoTitle",
          "seoDescription",
          "reactionCounts",
          "viewer"
        ]
      },
      "PostStatus": {
        "type": "string",
        "enum": ["ARCHIVED", "DRAFT", "PUBLISHED"]
      },
...
      "PaginatedPostAdminSummaryDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostAdminSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": ["items", "meta"]
      },
...
    }
  }
}